Refactor eth personal signer
